+++ mylog/lib/src/file.dart
@@ -5,13 +5,14 @@ import 'utils.dart';
 
 abstract class LogFileController {
   final _dir = Directory('logs');
-  final _sizeLimit = 1024 * 5;
+  int sizeLimit;
   late File _file;
   late String _fileName;
   // late IOSink _sink;
   final _linesQueue = StreamController<String>();
 
-  LogFileController() {
+  LogFileController({this.sizeLimit = 0, bool oneDayOneFile = true}) {
+    // FIXME: oneDayOneFile not finished
     _init();
   }
 
@@ -23,16 +24,16 @@ abstract class LogFileController {
     _fileName = DateTime.now().date;
     _file = File('logs/$_fileName.log');
     _linesQueue.stream.listen((event) async => await _logFile(event));
-    _checkDirSize();
   }
 
   Future<void> _logFile(String row) async {
     await _file.writeAsString(row, mode: FileMode.writeOnlyAppend);
-    _checkFileDate();
+    _nameControl();
+    _sizeControl();
   }
 
-  /// Check new date to change [_file]
-  void _checkFileDate() {
+  /// Check new date to create [_file]
+  void _nameControl() {
     final _today = DateTime.now().date;
     if (_fileName != _today) {
       _fileName = _today;
@@ -40,23 +41,28 @@ abstract class LogFileController {
     }
   }
 
-  /// Check [_dir] for file amount or size limit
-  void _checkDirSize() async {
+  /// Check [_dir] for [sizeLimit] and control free space
+  void _sizeControl() async {
+    if (sizeLimit == 0) return;
     var _total = 0;
     var _names = <String>[];
     try {
       var dirList = _dir.list();
-      await for (FileSystemEntity f in dirList) {
-        if (f is File) {
-          _total += await f.length();
-          _names.add(f.path.split(r'\').last.split('.').first);
+      // Check dir for log files
+      await for (FileSystemEntity _file in dirList) {
+        if (_file is File) {
+          _total += await _file.length();
+          _names.add(_file.path.split(r'\').last.split('.').first);
         }
       }
     } catch (e) {
       print(e.toString());
     }
-    if (_total > _sizeLimit) {
-      print('Log files size limit reached - $_total of $_sizeLimit.');
+    // TODO: file rotation for one file
+    if (_names.isEmpty || _names.length == 1) return;
+    // Check size
+    if (_total > sizeLimit) {
+      print('Log files size limit reached - $_total of $sizeLimit.');
       _names.sort();
       var _toRemove = File('logs/${_names.first}.log');
       if (await _toRemove.exists()) {