+++ mylog/lib/src/logger.dart
@@ -3,8 +3,10 @@ import 'dart:io';
 import 'file.dart';
 import 'utils.dart';
 
+enum Level { DEBUG, INFO, WARNING, ERROR, CRITICAL }
+
 class Logging extends LogFileController {
-  Level _level = Level.DEBUG;
+  Level lvl = Level.DEBUG;
 
   static final Logging _singleton = Logging._internal();
   factory Logging() => _singleton;
@@ -17,27 +19,24 @@ class Logging extends LogFileController {
   void e(String msg) => _log(msg, Level.ERROR);
   void c(String msg) => _log(msg, Level.CRITICAL);
 
-  void setLvl(Level lvl) => _level = lvl;
-  set lvl(Level lvl) => _level = lvl;
-
   /// Build log message, stdout/stderr it and put in [_msgQueue] (add middleware as listeners to this controller)
-  void _log(String msg, Level lvl) {
+  void _log(String msg, Level level) {
     var _time = DateTime.now().full;
     String _line; // line for file
-    if (lvl.index < _level.index) return;
-    if (lvl == Level.DEBUG) {
+    if (level.index < lvl.index) return;
+    if (level == Level.DEBUG) {
       _line = '$_time [VERB] $msg\n';
       stdout.write(_line);
-    } else if (lvl == Level.INFO) {
+    } else if (level == Level.INFO) {
       _line = '$_time [INFO] $msg\n';
       stdout.write(_line);
-    } else if (lvl == Level.WARNING) {
+    } else if (level == Level.WARNING) {
       _line = '$_time [WARN] $msg\n';
       stderr.write(_line);
-    } else if (lvl == Level.ERROR) {
+    } else if (level == Level.ERROR) {
       _line = '$_time [ERRO] $msg\n';
       stderr.write(_line);
-    } else if (lvl == Level.CRITICAL) {
+    } else if (level == Level.CRITICAL) {
       _line = '$_time [CRIT] $msg\n';
       stderr.write(_line);
     } else {